VK_KHR_variable_pointers: Don't use OpPtrAccessChain on a bare SSBO
You can't use OpPtrAccessChain on the direct result of an OpVariable
because it's not the element of any array. Instead, we need to do our
OpPtrAccessChain on a pointer to an actual array element and that
pointer needs to be properly decorated with a stride.  This is an
extension of CL #1864 which fixed this problem for the graphics tests
but not for compute.

Component: Vulkan

VK-GL-CTS issue: https://gitlab.khronos.org/Tracker/vk-gl-cts/issues/820

Affects:
dEQP-VK.spirv_assembly.instruction.compute.variable_pointers.complex_types_compute.opptraccesschain_*

Change-Id: Iab24d3d349f12bf015599277bd0d3836ca27fec3